renamed dbmodel to diagram
