+++ fennec/apps/diagram/admin.py
