+++ fennec/apps/versioncontroll/views.py
@@ -6,7 +6,7 @@ from django.contrib.auth.models import User, Group
 
 from fennec.apps.constants import MASTER_BRANCH_NAME, MASTER_BRANCH_TYPE, MASTER_BRANCH_DESCRIPTION
 
-from fennec.apps.dbmodel.serializers import SchemaSerializer, DiagramSerializer, SandboxBasicInfoSerializer
+from fennec.apps.diagram.serializers import SchemaSerializer, DiagramSerializer, SandboxBasicInfoSerializer
 from fennec.apps.versioncontroll import utils
 from fennec.apps.versioncontroll.models import Project, Branch, BranchRevision, SandboxChange
 from fennec.apps.versioncontroll.serializers import BranchRevisionSerializer, ChangeSerializer
